// frontend/src/app/services/models/page-response-dto-tool-transaction-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ToolTransactionDto } from '../models/tool-transaction-dto';
export interface PageResponseDtoToolTransactionDto {
  content?: Array<ToolTransactionDto>;
  first?: boolean;
  last?: boolean;
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}